moves ScanRequest/Response impls to their own module
