+++ zebra-grpc/src/methods.rs
@@ -1,0 +1,4 @@
+//! Types and functions used by RPC methods
+
+/// Implements RPC request/response types
+pub mod types;
